[react-gtk-rs] Initial commit
Add ReactLinux folder with base infrastructure.

Setup react-gtk-rs as a library and uiexplorer as a Gtk Rust application with
simpliest example possible which demonstrates JavaScriptCore C API interop.

So far, UIExplorer gets Segmentation fault under the pressure,
all pointes are opaque and some of the JSCS APIs are not covered yet.

This is also not tested on Linux environment.